Prevents management command form deleting all records if model has another manager called old_records that is not from the package

+++ django_old_records/management/commands/delete_old_records.py
@@ -1,4 +1,5 @@
 from django.core.management.base import BaseCommand
+from django_old_records.old_records_manager import OldRecordsManager
 from django.apps import apps
 
 class Command(BaseCommand):
@@ -7,6 +8,6 @@ class Command(BaseCommand):
     def handle(self, *args, **kwargs):
         models = apps.get_models()
         for model in models:
-            if hasattr(model, 'old_records'):
+            if hasattr(model, 'old_records') and isinstance(model.old_records, OldRecordsManager):
                 self.stdout.write(f'Deleting {model} old records')
                 model.old_records.all().delete()
